den Constructor auch geaendert

+++ SWE2_Micro_ERP_Schnittstelle/src/at/technikum/wien/winterhalder/kreuzriegler/swe2/request/GetContactByIdRequest.java
@@ -13,7 +13,7 @@ public class GetContactByIdRequest {
 	/**
 	 * @param contactId
 	 */
-	public GetContactByIdRequest(int contactId) {
+	public GetContactByIdRequest(long contactId) {
 		this.contactId = contactId;
 	}
